pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "shellagent"
version = "0.1.0"
description = "环境感知的智能Shell执行代理"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "openai>=1.30.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "distro>=1.9.0",
]

[project.scripts]
shellagent = "src.cli.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
